updated Cross-section values

+++ python/DYToMuMu_M_50To130_Tune4C_7TeV_pythia8_cff.py
@@ -1,7 +1,7 @@
 import FWCore.ParameterSet.Config as cms
 
 generator = cms.EDFilter("Pythia8GeneratorFilter",
-   crossSection = cms.untracked.double(770),
+   crossSection = cms.untracked.double(760),
    maxEventsToPrint = cms.untracked.int32(0),
    pythiaPylistVerbosity = cms.untracked.int32(1),
    filterEfficiency = cms.untracked.double(1.0),
